Upstream version 1.10.0

+++ tests/system_test.py
@@ -45,11 +45,27 @@ from threading import Event
 import json
 import uuid
 
+is_python2 = sys.version_info[0] == 2
+
+# DISPATCH-1443: for python < 2.7 use unittest2 since the default unittest for
+# older versions lacks features we need:
+#
+if is_python2 and sys.version_info[1] < 7:
+    # python < 2.7:
+    try:
+        import unittest2 as unittest
+    except ImportError:
+        raise Exception("Python unittest2 not installed - see README")
+else:
+    import unittest
+
 import proton
 from proton import Message
+from proton import Delivery
 from proton.handlers import MessagingHandler
 from proton.utils import BlockingConnection
 from proton.reactor import AtLeastOnce, Container
+from proton.reactor import AtMostOnce
 from qpid_dispatch.management.client import Node
 from qpid_dispatch_internal.compat import dict_iteritems
 
@@ -158,15 +174,16 @@ def get_local_host_socket(protocol_family='IPv4'):
 def port_available(port, protocol_family='IPv4'):
     """Return true if connecting to host:port gives 'connection refused'."""
     s, host = get_local_host_socket(protocol_family)
-
+    available = False
     try:
         s.connect((host, port))
-        s.close()
     except socket.error as e:
-        return e.errno == errno.ECONNREFUSED
+        available = e.errno == errno.ECONNREFUSED
     except:
         pass
-    return False
+
+    s.close()
+    return available
 
 def wait_port(port, protocol_family='IPv4', **retry_kwargs):
     """Wait up to timeout for port (on host) to be connectable.
@@ -175,14 +192,22 @@ def wait_port(port, protocol_family='IPv4', **retry_kwargs):
         """Only retry on connection refused"""
         if not isinstance(e, socket.error) or not e.errno == errno.ECONNREFUSED:
             raise
-    s, host = get_local_host_socket(protocol_family)
-    try:
-        retry_exception(lambda: s.connect((host, port)), exception_test=check,
-                        **retry_kwargs)
-    except Exception as e:
-        raise Exception("wait_port timeout on host %s port %s: %s"%(host, port, e))
 
-    finally: s.close()
+    host = None
+
+    def connect():
+        # macOS gives EINVAL for all connection attempts after a ECONNREFUSED
+        # man 3 connect: "If connect() fails, the state of the socket is unspecified. [...]"
+        s, host = get_local_host_socket(protocol_family)
+        try:
+            s.connect((host, port))
+        finally:
+            s.close()
+
+    try:
+        retry_exception(connect, exception_test=check, **retry_kwargs)
+    except Exception as e:
+        raise Exception("wait_port timeout on host %s port %s: %s" % (host, port, e))
 
 def wait_ports(ports, **retry_kwargs):
     """Wait up to timeout for all ports (on host) to be connectable.
@@ -243,7 +268,7 @@ class Process(subprocess.Popen):
                                 (args, kwargs, type(e).__name__, e))
 
     def assert_running(self):
-        """Assert that the proces is still running"""
+        """Assert that the process is still running"""
         assert self.poll() is None, "%s: exited" % ' '.join(self.args)
 
     def teardown(self):
@@ -267,6 +292,39 @@ class Process(subprocess.Popen):
             status = self.wait()
         if self.expect != None and self.expect != status:
             error("exit code %s, expected %s" % (status, self.expect))
+
+    def wait(self, timeout=None):
+        """
+        Add support for a timeout when using Python 2
+        """
+        if timeout is None:
+            return super(Process, self).wait()
+
+        if is_python2:
+            start = time.time()
+            while True:
+                rc = super(Process, self).poll()
+                if rc is not None:
+                    return rc
+                if time.time() - start >= timeout:
+                    raise Exception("Process did not terminate")
+                time.sleep(0.1)
+        else:
+            return super(Process, self).wait(timeout=timeout)
+
+    def communicate(self, input=None, timeout=None):
+        """
+        Add support for a timeout when using Python 2
+        """
+        if timeout is None:
+            return super(Process, self).communicate(input=input)
+
+        if is_python2:
+            self.wait(timeout=timeout)
+            return super(Process, self).communicate(input=input)
+
+        return super(Process, self).communicate(input=input,
+                                                timeout=timeout)
 
 
 class Config(object):
@@ -294,7 +352,7 @@ class Qdrouterd(Process):
             'listener': {'host':'0.0.0.0', 'saslMechanisms':'ANONYMOUS', 'idleTimeoutSeconds': '120',
                          'authenticatePeer': 'no', 'role': 'normal'},
             'connector': {'host':'127.0.0.1', 'saslMechanisms':'ANONYMOUS', 'idleTimeoutSeconds': '120'},
-            'router': {'mode': 'standalone', 'id': 'QDR', 'debugDumpFile': 'qddebug.txt'}
+            'router': {'mode': 'standalone', 'id': 'QDR'}
         }
 
         def sections(self, name):
@@ -340,6 +398,9 @@ class Qdrouterd(Process):
         self.perform_teardown = perform_teardown
         if not name: name = self.config.router_id
         assert name
+        # setup log and debug dump files
+        self.dumpfile = os.path.abspath('%s-qddebug.txt' % name)
+        self.config.sections('router')[0]['debugDumpFile'] = self.dumpfile
         default_log = [l for l in config if (l[0] == 'log' and l[1]['module'] == 'DEFAULT')]
         if not default_log:
             config.append(
@@ -374,6 +435,20 @@ class Qdrouterd(Process):
             return
 
         super(Qdrouterd, self).teardown()
+
+        # check router's debug dump file for anything interesting (should be
+        # empty) and dump it to stderr for perusal by organic lifeforms
+        try:
+            if os.stat(self.dumpfile).st_size > 0:
+                with open(self.dumpfile) as f:
+                    sys.stderr.write("\nRouter %s debug dump file:\n" % self.config.router_id)
+                    sys.stderr.write(f.read())
+                    sys.stderr.flush()
+        except OSError:
+            # failed to open file.  This can happen when an individual test
+            # spawns a temporary router (i.e. not created as part of the
+            # TestCase setUpClass method) that gets cleaned up by the test.
+            pass
 
     @property
     def ports_family(self):
@@ -526,17 +601,19 @@ class Tester(object):
         """
         @param id: module.class.method or False if no directory should be created
         """
-        self.directory = os.path.join(self.root_dir, *id.split('.'))
+        self.directory = os.path.join(self.root_dir, *id.split('.')) if id else None
         self.cleanup_list = []
 
     def rmtree(self):
         """Remove old test class results directory"""
-        shutil.rmtree(os.path.dirname(self.directory), ignore_errors=True)
+        if self.directory:
+            shutil.rmtree(os.path.dirname(self.directory), ignore_errors=True)
 
     def setup(self):
         """Called from test setup and class setup."""
-        os.makedirs(self.directory)
-        os.chdir(self.directory)
+        if self.directory:
+            os.makedirs(self.directory)
+            os.chdir(self.directory)
 
     def teardown(self):
         """Clean up (tear-down, stop or close) objects recorded via cleanup()"""
@@ -599,6 +676,7 @@ class TestCase(unittest.TestCase, Tester): # pylint: disable=too-many-public-met
 
     @classmethod
     def setUpClass(cls):
+        cls.maxDiff = None
         cls.tester = Tester('.'.join([cls.__module__, cls.__name__, 'setUpClass']))
         cls.tester.rmtree()
         cls.tester.setup()
@@ -730,8 +808,8 @@ class AsyncTestReceiver(MessagingHandler):
     Empty = Queue.Empty
 
     def __init__(self, address, source, conn_args=None, container_id=None,
-                 wait=True, recover_link=False):
-        super(AsyncTestReceiver, self).__init__()
+                 wait=True, recover_link=False, msg_args={}):
+        super(AsyncTestReceiver, self).__init__(**msg_args)
         self.address = address
         self.source = source
         self.conn_args = conn_args
@@ -798,16 +876,31 @@ class AsyncTestSender(MessagingHandler):
     A simple sender that runs in the background and sends 'count' messages to a
     given target.
     """
-    def __init__(self, address, target, count=1, body=None, container_id=None):
-        super(AsyncTestSender, self).__init__()
+    class TestSenderException(Exception):
+        def __init__(self, error=None):
+            super(AsyncTestSender.TestSenderException, self).__init__(error)
+
+    def __init__(self, address, target, count=1, message=None,
+                 container_id=None, presettle=False):
+        super(AsyncTestSender, self).__init__(auto_accept=False,
+                                              auto_settle=False)
         self.address = address
         self.target = target
-        self.count = count
-        self._unaccepted = count
-        self._body = body or "test"
+        self.total = count
+        self.presettle = presettle
+        self.accepted = 0
+        self.released = 0
+        self.modified = 0
+        self.rejected = 0
+        self.sent = 0
+        self.error = None
+        self.link_stats = None
+
+        self._message = message or Message(body="test")
         self._container = Container(self)
         cid = container_id or "ATS-%s:%s" % (target, uuid.uuid4())
         self._container.container_id = cid
+        self._link_name = "%s-%s" % (cid, "tx")
         self._thread = Thread(target=self._main)
         self._thread.daemon = True
         self._thread.start()
@@ -815,41 +908,75 @@ class AsyncTestSender(MessagingHandler):
     def _main(self):
         self._container.start()
         while self._container.process():
-            pass
+            self._check_if_done()
 
     def wait(self):
         # don't stop it - wait until everything is sent
         self._thread.join(timeout=TIMEOUT)
         assert not self._thread.is_alive(), "sender did not complete"
+        if self.error:
+            raise AsyncTestSender.TestSenderException(self.error)
 
     def on_start(self, event):
         self._conn = self._container.connect(self.address)
 
     def on_connection_opened(self, event):
+        option = AtMostOnce if self.presettle else AtLeastOnce
         self._sender = self._container.create_sender(self._conn,
                                                      target=self.target,
-                                                     options=AtLeastOnce())
+                                                     options=option(),
+                                                     name=self._link_name)
 
     def on_sendable(self, event):
-        if self.count:
-            self._sender.send(Message(body=self._body))
-            self.count -= 1
+        if self.sent < self.total:
+            self._sender.send(self._message)
+            self.sent += 1
 
-    def on_accepted(self, event):
-        self._unaccepted -= 1;
-        if self._unaccepted == 0:
+    def _check_if_done(self):
+        done = (self.sent == self.total
+                and (self.presettle
+                     or (self.accepted + self.released + self.modified
+                         + self.rejected == self.sent)))
+        if done and self._conn:
+            self.link_stats = get_link_info(self._link_name,
+                                            self.address)
             self._conn.close()
             self._conn = None
+
+    def on_accepted(self, event):
+        self.accepted += 1;
+        event.delivery.settle()
+
+    def on_released(self, event):
+        # for some reason Proton 'helpfully' calls on_released even though the
+        # delivery state is actually MODIFIED
+        if event.delivery.remote_state == Delivery.MODIFIED:
+            return self.on_modified(event)
+        self.released += 1
+        event.delivery.settle()
+
+    def on_modified(self, event):
+        self.modified += 1
+        event.delivery.settle()
+
+    def on_rejected(self, event):
+        self.rejected += 1
+        event.delivery.settle()
+
+    def on_link_error(self, event):
+        self.error = "link error:%s" % str(event.link.remote_condition)
+        self._conn.close()
+        self._conn = None
 
 
 class QdManager(object):
     """
     A means to invoke qdmanage during a testcase
     """
-    def __init__(self, tester, address=None, timeout=TIMEOUT):
+    def __init__(self, tester=None, address=None, timeout=TIMEOUT):
         # 'tester' - can be 'self' when called in a test,
         # or an instance any class derived from Process (like Qdrouterd)
-        self._tester = tester
+        self._tester = tester or Tester(None)
         self._timeout = timeout
         self._address = address
 
@@ -872,6 +999,16 @@ class QdManager(object):
 
     def create(self, long_type, kwargs):
         cmd = "CREATE --type=%s" % long_type
+        for k, v in kwargs.items():
+            cmd += " %s=%s" % (k, v)
+        return json.loads(self(cmd))
+
+    def update(self, long_type, kwargs, name=None, identity=None):
+        cmd = 'UPDATE --type=%s' % long_type
+        if identity is not None:
+            cmd += " --identity=%s" % identity
+        elif name is not None:
+            cmd += " --name=%s" % name
         for k, v in kwargs.items():
             cmd += " %s=%s" % (k, v)
         return json.loads(self(cmd))
@@ -926,6 +1063,10 @@ class MgmtMsgProxy(object):
     def response(self, msg):
         ap = msg.properties
         return self._Response(ap['statusCode'], ap['statusDescription'], msg.body)
+
+    def query_router(self):
+        ap = {'operation': 'QUERY', 'type': 'org.apache.qpid.dispatch.router'}
+        return Message(properties=ap, reply_to=self.reply_addr)
 
     def query_connections(self):
         ap = {'operation': 'QUERY', 'type': 'org.apache.qpid.dispatch.connection'}
@@ -1006,3 +1147,27 @@ class TestTimeout(object):
 
     def on_timer_task(self, event):
         self.parent.timeout()
+
+
+class PollTimeout(object):
+    """
+    A callback object for MessagingHandler scheduled timers
+    parent: A MessagingHandler with a poll_timeout() method
+    """
+    def __init__(self, parent):
+        self.parent = parent
+
+    def on_timer_task(self, event):
+        self.parent.poll_timeout()
+
+
+def get_link_info(name, address):
+    """
+    Query the router at address for the status and statistics of the named link
+    """
+    qdm = QdManager(address=address)
+    rc = qdm.query('org.apache.qpid.dispatch.router.link')
+    for item in rc:
+        if item.get('name') == name:
+            return item
+    return None
